refac: useDialogContext를 이용한 모달, 알림 변경
