+++ src/components/announcement/layout/AnnouncementSideView.tsx
@@ -1,11 +1,27 @@
+import { useCalendarData } from '@/hooks';
+
 import Tag from '@/components/common/Tag';
 import Title from '@/components/common/Title';
+import Calendar from '@/components/schedule/Calendar';
+import SmallCalendarBottomEvent from '@/components/schedule/SmallCalendarBottomEvent';
 
 export default function AnnouncementSideView() {
+  const { fullCalendarEvents } = useCalendarData();
+
   return (
     <>
       <Title title="주요일정" />
-      <div className="mb-10 mt-2 h-60 rounded-xl border bg-white shadow-card" />
+      <Calendar type="small" events={fullCalendarEvents}>
+        <ul className="mb-1 flex flex-col gap-2">
+          {fullCalendarEvents.map(event => (
+            <SmallCalendarBottomEvent
+              key={event.title}
+              date={new Date(event.start).toLocaleDateString()}
+              title={event.title}
+            />
+          ))}
+        </ul>
+      </Calendar>
 
       <div className="mb-2 flex items-center justify-between">
         <Title title="최근 본 공지사항" className="!pl-0 text-sm" />
