+++ src/components/common/button/VerifyCodeButton.tsx
@@ -1,7 +1,6 @@
-import { useToggleModal } from '@/hooks';
+import { useDialogContext } from '@/hooks';
 
 import SquareButton from '@/components/common/button/SquareButton';
-import Alert from '@/components/common/modal/Alert';
 
 const CODE = '새싹';
 
@@ -12,38 +11,34 @@ interface VerifyCodeButtonProps {
 export default function VerifyCodeButton({
   currentCode,
 }: VerifyCodeButtonProps) {
-  const { modalOpen, toggleModal } = useToggleModal();
-
-  const onVerifyCodeClick = () => {
-    toggleModal();
-  };
+  const { alert, hideDialog } = useDialogContext();
 
   const verified = CODE === currentCode;
 
-  return (
-    <>
-      <button
-        type="button"
-        className="absolute right-0 top-12 mt-2 rounded-md border px-2 py-0.5 text-gray1"
-        onClick={onVerifyCodeClick}
-      >
-        인증 확인
-      </button>
+  const onVerifyCodeClick = () => {
+    alert({
+      showDim: true,
+      className: 'z-30',
+      text: verified ? '인증 확인되었습니다!' : '인증에 실패했습니다.',
+      subText: !verified ? '관리자에게 문의해주세요' : undefined,
+      children: (
+        <SquareButton
+          name="나가기"
+          onClick={hideDialog}
+          type="button"
+          className="mt-5"
+        />
+      ),
+    });
+  };
 
-      {modalOpen && (
-        <Alert
-          text={verified ? '인증 확인되었습니다!' : '인증에 실패했습니다.'}
-          subText={!verified ? '관리자에게 문의해주세요' : undefined}
-          className="gap-1"
-        >
-          <SquareButton
-            name="나가기"
-            onClick={toggleModal}
-            type="button"
-            className="mt-5"
-          />
-        </Alert>
-      )}
-    </>
+  return (
+    <button
+      type="button"
+      className="absolute right-0 top-12 mt-2 rounded-md border px-2 py-0.5 text-gray1"
+      onClick={onVerifyCodeClick}
+    >
+      인증 확인
+    </button>
   );
 }
